I think that does it, for now

+++ Depth2Depth/dishwasher_depth2depth_shared.py
@@ -258,7 +258,7 @@ class MyLRSchedule(tf.keras.optimizers.schedules.LearningRateSchedule):
         self.batch_size = batch_size
 
     def __call__(self, step):
-        return tf.cond((step / self.batch_size) < 100, lambda: self.initial_learning_rate, lambda: self.initial_learning_rate - ((step / self.batch_size) - 100) * (self.initial_learning_rate / 100))
+        return tf.cond((step / self.batch_size) < 100, lambda: tf.cast(self.initial_learning_rate, dtype=tf.float32), lambda: tf.cast(self.initial_learning_rate - ((step / self.batch_size) - 100) * (self.initial_learning_rate / 100), dtype=tf.float32))
 
 
 def get_networks(g_in_channels, g_out_channels, f_in_channels, f_out_channels):
